# TicketManagementFrontEnd/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SimSoftAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/problem-categories": {
      "get": {
        "tags": [
          "ProblemCategory"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ProblemCategory"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/problem-categories/{id}": {
      "delete": {
        "tags": [
          "ProblemCategory"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/problem-categories/assign-to-client": {
      "post": {
        "tags": [
          "ProblemCategory"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProblemCategoryToClientDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProblemCategoryToClientDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProblemCategoryToClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/problem-categories/client/{clientId}": {
      "get": {
        "tags": [
          "ProblemCategory"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignProblemCategoryToClientDto": {
        "required": [
          "clientId",
          "problemCategoryId"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "problemCategoryId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemCategory": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "parentCategoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "parentCategory": {
            "$ref": "#/components/schemas/ProblemCategory"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}